+++ lib/hexapdf/task.rb
@@ -64,6 +64,7 @@ module HexaPDF
 
     autoload(:Optimize, 'hexapdf/task/optimize')
     autoload(:Dereference, 'hexapdf/task/dereference')
+    autoload(:PDFA, 'hexapdf/task/pdfa')
 
   end